feat: created a grid for the toolcards and an expanded version of toolcard
Created a grid for the non expanded toolcards, and an expandable toolcard that expands when user presses expand button on toolcard. The expandable toolcard takes inspiration from an example in aceternity, and the added useOutside-click hook is entirely from aceternity except changing any types to a valid type in side the hook. The hook ensures the expanded card disappears when user clicks outside or on minimize button.