feat: implement calendar for reservations
Implemented FullCalendar for reservations in shadcn styling. Created a CustomToolBar and CustomEventStyling in order to control the styling, translations etc. on buttons, text etc. Added some comments for explanation in certain parts. There are many constants just for the purpose of testing the creation and editing of reservations on the calendar.