feat: create a reservationform and a dialogmodal for when creating events on the calendar
Created a form for when user reserves a tool and a dialog that pops up with the reservation form when user selects and drags a slot on the calendar, or when user presses create reservation button. Added DateTimePicker from shadcn as the current seperate date and time pickers posed challenges when formatting the meridian for time and order of days, month, year for date.